pyproject.toml
[tool.poetry]
name = "semantic-search-elasticsearch-openai-langchain"
version = "0.1.0"
description = ""
authors = ["Dylan Castillo <dylanjcastillo@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "semantic_search_elasticsearch_openai_langchain"}]

[tool.poetry.dependencies]
python = "^3.10"
openai = "^0.27.4"
langchain = "^0.0.147"
elasticsearch = "^8.7.0"
python-dotenv = "^1.0.0"
beautifulsoup4 = "^4.12.2"
lxml = "^4.9.2"
tiktoken = "^0.3.3"
fastapi = {extras = ["all"], version = "^0.95.1"}

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
flake8 = "^6.0.0"
mypy = "^1.2.0"
notebook = "^6.5.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
